trending-topic-collector/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
pyyaml>=6.0.0
pytest>=8.0.0
